requirements.txt
selenium>=4.8.0
webdriver-manager>=3.8.5
numpy>=1.22.0
opencv-python>=4.6.0
pillow>=9.3.0
scikit-image>=0.19.0
requests>=2.28.1
tqdm>=4.64.1